fix(core): Use cgpv-main path to access assets Closes #2123
