+++ packages/geoview-core/src/core/utils/utilities.ts
@@ -44,6 +44,34 @@ export function getLocalizedMessage(localizedKey: string, language: TypeDisplayL
 }
 
 /**
+ * Get the URL of main script cgpv-main so we can access the assets
+ * @returns {string} the URL of the main script
+ */
+export function getScriptAndAssetURL(): string {
+  // get all loaded js scripts on the page
+  const scripts = document.getElementsByTagName('script');
+  let scriptPath: string = '';
+
+  if (scripts && scripts.length) {
+    // go through all loaded scripts on the page
+    for (let scriptIndex = 0; scriptIndex < scripts.length; scriptIndex++) {
+      // search for the core script
+      if (scripts[scriptIndex].src.includes('cgpv-main')) {
+        // get the src of the core script
+        const { src } = scripts[scriptIndex];
+
+        // extract the host from the loaded core script
+        scriptPath = src.substring(0, src.lastIndexOf('/'));
+
+        break;
+      }
+    }
+  }
+
+  return scriptPath;
+}
+
+/**
  * Generate a unique id if an id was not provided
  * @param {string} id an id to return if it was already passed
  * @returns {string} the generated id
@@ -416,9 +444,13 @@ function getSectionHeading(content: string): string {
  * @param {TypeDisplayLanguage} language - Language to use for guide.
  * @returns {Promise<TypeGuideObject | undefined>} The guide object
  */
-export async function createGuideObject(mapId: string, language: TypeDisplayLanguage): Promise<TypeGuideObject | undefined> {
+export async function createGuideObject(
+  mapId: string,
+  language: TypeDisplayLanguage,
+  assetsURL: string
+): Promise<TypeGuideObject | undefined> {
   try {
-    const response = await fetch(`./locales/${language}/guide.md`);
+    const response = await fetch(`${assetsURL}/locales/${language}/guide.md`);
     const content = await response.text();
 
     // Split by first level sections (Split with =1!<key>=)
